hecho que descuente la venta virtual y guarde los movimientos

+++ DisplaBackend/DisplaBackend/Models/ArticuloVario.cs
@@ -15,8 +15,8 @@ namespace DisplaBackend.Models
         public int Id { get; set; }
         public string Nombre { get; set; }
         public int StockMinimo { get; set; }
-        public int StockActual { get; set; }
-        public double PrecioCosto { get; set; }
+        public decimal StockActual { get; set; }
+        public decimal PrecioCosto { get; set; }
         public double PorcentajeUtilidad { get; set; }
         public int IdTipoArticulo { get; set; }
         public bool Borrado { get; set; }
