+++ DisplaBackend/DisplaBackend/Models/ComprobanteItemLente.cs
@@ -8,10 +8,10 @@ namespace DisplaBackend.Models
         public int Id { get; set; }
         public int IdLente { get; set; }
         public int IdComprobanteItem { get; set; }
-        public double Cantidad { get; set; }
+        public decimal Cantidad { get; set; }
         public decimal? MedidaEsferico { get; set; }
         public decimal? MedidaCilindrico { get; set; }
-        public double Precio { get; set; }
+        public decimal Precio { get; set; }
 
         public virtual ComprobanteItem IdComprobanteItemNavigation { get; set; }
         public virtual Lente IdLenteNavigation { get; set; }
