+++ DisplaBackend/DisplaBackend/Models/StockLente.cs
@@ -9,7 +9,7 @@ namespace DisplaBackend.Models
         public decimal MedidaEsferico { get; set; }
         public decimal MedidaCilindrico { get; set; }
         public int IdLente { get; set; }
-        public double Stock { get; set; }
+        public decimal Stock { get; set; }
 
         public virtual Lente IdLenteNavigation { get; set; }
     }
